pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drug-safety-assistant"
version = "0.1.0"
description = "Evidence-grounded drug safety decision support assistant"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "requests>=2.31.0",
  "pydantic>=2.6.0",
  "PyYAML>=6.0.1",
  "python-dotenv>=1.0.1",
  "numpy>=1.24.0",
  "streamlit>=1.40.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
]
api = [
  "anthropic>=0.40.0",
]
ml = [
  "faiss-cpu>=1.8.0",
  "sentence-transformers>=3.0.0",
  "transformers>=4.46.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
